+++ src/julia-export/index.js
@@ -12,8 +12,6 @@ class JuliaExport extends _Export {
       this.spaceFilter = q.spaceFilter;
     } else if (typeof q.spaceFilter === 'string') {
       this.spaceFilter = [q.spaceFilter];
-    } else {
-      this.spaceFilter = ['nameless'];
     }
 
     return this;
@@ -22,29 +20,24 @@ class JuliaExport extends _Export {
     return 'JuliaExport';
   }
   make(){
-    // use only one namespace
     let logger = this.container.logger;
-    if (this.spaceFilter.length === 0) {
-      let msg = 'spaceFilter for Julia format should include at least one namespace but get empty';
-      logger.err(msg);
-      var modelContent = '';
-      var runContent = '';
-    } else if (!this.container.namespaces.has(this.spaceFilter[0])) {
-      let msg = `Namespace "${this.spaceFilter[0]}" does not exist.`;
-      logger.err(msg);
-      modelContent = '';
-      runContent = '';
-    } else {
-      if (this.spaceFilter.length > 1) {
-        let msg = `Julia format does not support multispace export. Only first namespace "${this.spaceFilter[0]}" will be used.`;
-        logger.warn(msg);
-      }
-      let ns = this.container.namespaces.get(this.spaceFilter[0]);
-     
-      let image = this.getJuliaImage(ns);
-      modelContent = this.getModelCode(image);
-      runContent = this.getRunCode(image);
-    }
+    // create image for multiple namespaces
+    let nsImages = [...this.container.namespaces]
+      .filter((pair) => {
+        let allowedByFilter = typeof this.spaceFilter === 'undefined'
+          || this.spaceFilter.indexOf(pair[0]) !== -1;
+        return allowedByFilter && !pair[1].isAbstract;
+      })
+      .map((pair) => this.getJuliaImage(pair[1]));
+
+    // create Content
+    let image = {
+      builderName: pkg.name + ' of v' + pkg.version,
+      options: this,
+      nsImages
+    };
+    let modelContent = this.getModelCode(image);
+    let runContent = this.getRunCode(image);
 
     return [
       {
@@ -60,8 +53,6 @@ class JuliaExport extends _Export {
     ];
   }
   getJuliaImage(ns){
-    let builderName = pkg.name + ' of v' + pkg.version;
-    
     // constants
     let constants = ns
       .selectByInstanceOf('Const');
@@ -126,9 +117,7 @@ class JuliaExport extends _Export {
     let pTranslatorArray = constants
       .map((x, i) => [x.id, `cons[${i+1}]`]);
       
-    return { 
-      builderName,
-      options: this,
+    return {
       namespace: ns,
       constants,
       dynamicRecords,
@@ -141,13 +130,13 @@ class JuliaExport extends _Export {
       pTranslator: { symbolName: _.fromPairs(pTranslatorArray)},
     };
   }
-  getModelCode(image = {}){
+  getModelCode(image = []){
     return nunjucks.render(
       'model.jl.njk',
       image
     );
   }
-  getRunCode(image = {}){
+  getRunCode(image = []){
     return nunjucks.render(
       'run.jl.njk',
       image